# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: EventHub API
  version: "1.0.0"
  description: "Documentation \"EventHub\" API v1.0"
servers:
  - url: http://localhost:8080
...
paths:
...
  /events/v1/api:
    post:
      tags:
        - event
      summary: Создать событие
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateDto'
      responses:
        '201':
          description: Событие успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - event
      summary: Список событий
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
          required: false
          description: "Фильтр: вернуть события только указанного пользователя"
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedEventDtoResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/v1/api/{id}:
    get:
      tags:
        - event
      summary: Получить событие по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Событие найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDto'
        '400':
          description: Некорректный идентификатор
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - event
      summary: Обновить событие
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdateDto'
      responses:
        '200':
          description: Событие обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - event
      summary: Удалить событие
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Событие удалено
        '400':
          description: Некорректный идентификатор
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/v1/api/tags:
    post:
      tags:
        - tag
      summary: Создать тег
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateDto'
      responses:
        '201':
          description: Тег успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Тег с таким именем уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - tag
      summary: Получить список тегов
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Список тегов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTagDtoResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/v1/api/tags/{id}:
    get:
      tags:
        - tag
      summary: Получить тег по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Тег найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagDto'
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тег не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - tag
      summary: Обновить тег
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdateDto'
      responses:
        '200':
          description: Тег обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagDto'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тег не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Тег с таким именем уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - tag
      summary: Удалить тег
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Тег удалён
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тег не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    EventCreateDto:
      type: object
      required: [ title, createdBy, categoryId ]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
          description: Название события
          example: "Митап по Java"
        description:
          type: string
          maxLength: 255
          description: Краткое описание события
          example: "Встреча для обсуждения новых технологий Java."
        eventDate:
          type: string
          format: date-time
          description: Дата и время проведения
          example: "2025-12-10T18:00:00+03:00"
        location:
          type: string
          description: Место проведения
          example: "Москва, ул. Ленина, 10"
        categoryId:
          type: string
          format: uuid
          description: Идентификатор категории
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        createdBy:
          type: string
          format: uuid
          description: Идентификатор пользователя, создавшего событие
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        tags:
          type: array
          description: Список идентификаторов тегов
          items:
            type: string
            format: uuid
            example: "c1d2e3f4-5678-1234-abcd-9876543210ab"
    EventUpdateDto:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          description: Новое название события
          example: "Spring Boot Workshop"
        description:
          type: string
          maxLength: 255
          description: Новое описание события
          example: "Практический семинар по Spring Boot."
        eventDate:
          type: string
          format: date-time
          description: Новая дата и время проведения
          example: "2025-12-15T19:00:00+03:00"
        location:
          type: string
          description: Новое место проведения
          example: "Санкт-Петербург, пр. Невский, 100"
        categoryId:
          type: string
          format: uuid
          description: Новый идентификатор категории
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        createdBy:
          type: string
          format: uuid
          description: Новый идентификатор пользователя
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        tags:
          type: array
          description: Новый список идентификаторов тегов
          items:
            type: string
            format: uuid
            example: "c1d2e3f4-5678-1234-abcd-9876543210ab"
        tagUpdateMode:
          type: string
          enum: [ REPLACE, ADD, REMOVE ]
          description: Режим обновления тегов
          example: "REPLACE"
...
    TagCreateDto:
      type: object
      required: [ name ]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: Название тега
          example: "Java"
        description:
          type: string
          maxLength: 255
          description: Описание тега
          example: "Тег для событий, связанных с Java."
    TagUpdateDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          description: Новое название тега
          example: "Spring"
        description:
          type: string
          maxLength: 255
          description: Новое описание тега
          example: "Тег для событий по Spring Framework."
...
    EventDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор события
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        title:
          type: string
          description: Название события
          example: "Митап по Java"
        description:
          type: string
          description: Краткое описание события
          example: "Встреча для обсуждения новых технологий Java."
        eventDate:
          type: string
          format: date-time
          description: Дата и время проведения
          example: "2025-12-10T18:00:00+03:00"
        location:
          type: string
          description: Место проведения
          example: "Москва, ул. Ленина, 10"
        category:
          $ref: '#/components/schemas/CategoryShortDto'
        createdBy:
          $ref: '#/components/schemas/UserShortDto'
        tags:
          type: array
          description: Список тегов события
          items:
            $ref: '#/components/schemas/TagShortDto'
    TagShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор тега
          example: "c1d2e3f4-5678-1234-abcd-9876543210ab"
        name:
          type: string
          description: Название тега
          example: "Java"
    UserShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор пользователя
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        username:
          type: string
          description: Имя пользователя
          example: "ivan.petrov"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "ivan.petrov@example.com"
    CategoryShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор категории
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        name:
          type: string
          description: Название категории
          example: "Технологии"
    TagDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор тега
          example: "c1d2e3f4-5678-1234-abcd-9876543210ab"
        name:
          type: string
          description: Название тега
          example: "Java"
        description:
          type: string
          description: Описание тега
          example: "Тег для событий, связанных с Java."
        events:
          type: array
          description: Список событий с этим тегом
          items:
            $ref: '#/components/schemas/EventShortDto'
    EventShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор события
          example: "b1a7e8c2-3d4f-4e2a-9c1a-2b3c4d5e6f7a"
        title:
          type: string
          description: Название события
          example: "Митап по Java"
...
    ErrorResponse:
      type: object
      properties:
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: "/events"
        httpMethod:
          type: string
          description: HTTP метод запроса
          example: "POST"
        statusCode:
          type: integer
          description: Код ошибки
          example: 400
        errorReason:
          type: string
          description: Краткое описание причины ошибки
          example: "Bad Request"
        message:
          type: string
          description: Описание ошибки
          example: "Некорректные данные запроса"
    PagedEventDtoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EventDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    PagedTagDtoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TagDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer